example: reuse the shared player on Android
Replace the nested mobile module and duplicated player UI with a thin gomobile adapter that registers the shared game and preserves the optional document-picker bridge.

Keep Android packaging self-contained by preparing the pinned go-ffmpeg-ffi libraries from the Makefile while retaining FFMPEG_ANDROID_LIB_DIR for CI and prebuilt outputs.
