+++ examples/mediaplayer/android/.gitignore
