--- examples/mobileplayer/go.mod
@@ -1,21 +1,0 @@
-module github.com/erparts/go-avebi/examples/mobileplayer
-
-go 1.25.0
-
-require (
-	github.com/erparts/go-avebi v0.0.0
-	github.com/hajimehoshi/ebiten/v2 v2.9.9
-)
-
-require (
-	github.com/bstkhq/go-ffmpeg-ffi v1.1.0 // indirect
-	github.com/ebitengine/gomobile v0.0.0-20250923094054-ea854a63cce1 // indirect
-	github.com/ebitengine/hideconsole v1.0.0 // indirect
-	github.com/ebitengine/oto/v3 v3.4.0 // indirect
-	github.com/ebitengine/purego v0.9.1 // indirect
-	github.com/jezek/xgb v1.1.1 // indirect
-	golang.org/x/sync v0.20.0 // indirect
-	golang.org/x/sys v0.41.0 // indirect
-)
-
-replace github.com/erparts/go-avebi => ../..
